fdt: Move all fdt files into the same directory
We want to build device tree source files from /u-boot/dts, so we must
make them build both inside and outside the U-Boot tree. This requires that
we remove the relative paths from the front of *.dts files.

Also we should keep the convention that include files end with .dtsi

It would be more convenient to add a -I option to dtc to specify a list
of include directories to search.

BUG=chromium-os:17192
TEST=emerge U-Boot, emerge tegra2-fdt, see that both succeed

Change-Id: I2ac65296cd19ee92e02837c9b418c8a630ef3075
Reviewed-on: http://gerrit.chromium.org/gerrit/3667
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>