CHROMIUM: Move CONFIG_CROSSYSTEM_DATA_ADDRESS into a header file.
This flag is only valid in old vboot api. The new vboot api logic puts
crossystem data in a global data struct. But without declaring this flag
in chromeos_seaboard_vboot config cause the build failed. So we need to
move this flag out of config and declare it in a header file.

BUG=none
TEST=build chromeos_seaboard_vboot without error

Change-Id: Ice26bb477c165be4a28848d7d13bf5d60e414c1d
Reviewed-on: http://gerrit.chromium.org/gerrit/3840
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>