Make the software cursor in the graphical framebuffer non-destructive.
When printing the string "\r\n" to the framebuffer console, the first character
of the current line was being replaced with a space. The "boot" prompt would
become the "oot" prompt. This change makes the cursor non-destructive so that
no matter where it goes on its way to where it's supposed to be, the end result
is that the cursor is where it's supposed to be with the other text preserved
intact.

BUG=chrome-os-partner:4968
TEST=Built and booted with a static and blinking software cursor. Entered
commands, scrolled the screen, used backspace, and verified that the cursor
behaved as expected. Verified that the "oot" prompt problem was gone.

Change-Id: I7ba6b20b7c9131f904057adbe4cc033bda04f1bb
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4081
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>