Add a CBFS driver and commands to u-boot.
This change adds CBFS support and some commands to use it to u-boot. These
commands are:

cbfsinit - Initialize CBFS support and pull all metadata into RAM. The end of
the ROM is an optional parameter which defaults to the standard 0xffffffff and
can be used to support multiple CBFSes in a system. The last one set up with
cbfsinit is the one that will be used.

cbfsinfo - Print information from the CBFS header.

cbfsls - Print out the size, type, and name of all the files in the current
CBFS. Recognized types are translated into symbolic names.

cbfsload - Load a file from CBFS into memory. Like the similar command for fat
filesystems, you can optionally provide a maximum size.

Support for CBFS is compiled in when the CONFIG_CMD_CBFS option is specified.

BUG=chrome-os-partner:3910
TEST=Built and booted on an Alex. Initialized with and without specifying the
end of the ROM, and with a bad end of ROM. Ran the commands before CBFS was
initialized. Ran cbfsinfo and saw reasonable output. Ran cbfsls and saw output
that matched what was printed when the CBFS was put together by the coreboot
ebuild. Used cbfsload to load a test text file into memory and verified that it
was the correct size and had the correct contents. Ran with a max size and saw
the file was truncated in memory.

Change-Id: I64d06d49633cef3cffac1d571519eae38c7d267f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4167
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>