No longer build build a separate device tree as part of u-boot.
This changes DEV_TREE_SEPARATE so that it just means that u-boot should
expect the dev tree to be separate, not that it should build it.
We're already building separate device trees as part of the FDT build,
and we're already using those device trees instead of this one.

BUG=chromium-os:17191
TEST=Ran emerge-${BOARD} chromeos-u-boot chromeos-bootimage
Flashed u-boot onto the board w/ cros_write_firmware

Change-Id: I948e96b2b2b15255c5b0a6244c7bbc1972e139c3
Reviewed-on: http://gerrit.chromium.org/gerrit/4267
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Anton Staaf <robotboy@chromium.org>