CHROMIUM: ARM: tegra: Don't run 'regen_all' in secure boot case.
This requires 'bootargs' to be correct by default, so I did
a little bit of refactoring of the way regen_all works to
account for that.

After this CL lands and is marked as stable, I'll change the
FDT from 'run secure_boot' to just 'vboot_twostop'.  That
will mean that we're not _running_ any commands from the
environment in the vboot case.

BUG=chromium-os:17940
TEST=Saw the legacy boot worked; saw that vboot worked.

Change-Id: I966c3f215beef5d9a1e47fa483a89f0ec42b8914
Reviewed-on: http://gerrit.chromium.org/gerrit/4448
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>